refac: 예약 기간 타입 long -> int로 변경

+++ src/test/java/com/prgrms/amabnb/reservation/entity/vo/ReservationDateTest.java
@@ -94,13 +94,13 @@ class ReservationDateTest {
     @Test
     void getPeriod() {
         // given
-        long period = 5;
+        int period = 5;
         LocalDate checkIn = LocalDate.now();
         LocalDate checkOut = checkIn.plusDays(5L);
         ReservationDate reservationDate = new ReservationDate(checkIn, checkOut);
 
         // when
-        long calculatePeriod = reservationDate.getPeriod();
+        int calculatePeriod = reservationDate.getPeriod();
 
         // then
         assertThat(calculatePeriod).isEqualTo(period);
